binance-tick-streaming/requirements.txt
faust-streaming==0.11.3
requests==2.32.3